+++ src/imports.rs
@@ -1,19 +1,17 @@
-use _macros;
-
 // Notice how we use self:: and use here to import into the local scope
 // so that hello is available.
 use self::imported::Imported;
 
 mod imported {
   pub trait Imported {
-    fn hello(&self, x:int) -> int;
+    fn hello(&self, x:isize) -> isize;
   }
 }
 
 struct Hi;
 
 impl imported::Imported for Hi {
-  fn hello(&self, x:int) -> int {
+  fn hello(&self, x:isize) -> isize {
     return x;
   }
 }
